Fix refactoring problems #2303

+++ libraries/plugins/account_history_rocksdb/include/steem/plugins/account_history_rocksdb/account_history_rocksdb_objects.hpp
@@ -36,8 +36,8 @@ class volatile_operation_object : public object< volatile_operation_object_type,
       chain::transaction_id_type trx_id;
       uint32_t                   block = 0;
       uint32_t                   trx_in_block = 0;
-      uint16_t                   op_in_trx = 0;
-      uint16_t                   virtual_op = 0;
+      uint32_t                   op_in_trx = 0;
+      uint32_t                   virtual_op = 0;
       time_point_sec             timestamp;
       chain::buffer_type         serialized_op;
       chainbase::t_vector< account_name_type > impacted;
